+++ billy/models/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import unicode_literals
 
 from base import Base
-from groups import Company
+from company import Company
 from coupons import Coupon
-from payouts import Payout
-from plans import ChargePlan
+from payout_plans import Payout
+from charge_plans import ChargePlan
 from customers import Customer
-from plan_invoice import PlanSubscription, PlanInvoice
+from charge_invoice import PlanSubscription, PlanInvoice
 from payout_invoice import PayoutSubscription, PayoutInvoice
 from transactions import PlanTransaction, PayoutTransaction
